Add files via upload
This is the same test sketch as V1, but here I use the millis() function instead of delay and use 2 as the Interrupt Pin to get a better reading for the HALL sensor. Using this, I could operate more trashcans from one Arduino. Using the delay() function reduces the scalability of the sketch to other trashcans. Because the Arduino Uno only has a few (2!) interrupt pins and no build-in WiFi, this project might be continued on a different microprocessor like the ESP32. On the ESP32 all the pins can be used as interrupt pins.