Add files via upload
In this script I have linked the PIR motion sensor to the Arduino. If Motion is Detected, the trashcan closes quickly. When no motion is detected afterwards, the trashcan opens in 15 seconds. If motion is detected, the timer resets and the time is 15 seconds again. The next step is too link multiple trashcan, but I need more magnets for this. I found some in my room from the previous tenant, but I lost them somehow. Maybe the stuck to something? Anyways I ordered 100 more so I'll continue on this next week, installing more Hall Effect sensors in more trashcans. I'll try using it with an ESP32 tomorrow.