Add files via upload
Removed the Interrupt and made an if loop during the CLOSING state. It gave some issues in some of the sensors for some reason. Doing this stabilized the behavior significantly. This also means I can scale the code to multiple Trashcans using the same set up before I port it to the ESP32. Small steps!