Create ESP32 Test One Trashcan (ShyTrashcansV1.ino)
This is an iteration on ShyTrashcansV1.ino code, but created to optimize for the ESP32. The code was uploaded to a ESP32 WROOM board and uploaded using Arduino IDE. This code tests whether the ESP32 reads the HALL sensor in the Trashcan and can control the motors. By using the ledcWrite() function, it is possible to get rid of the high pitched PWM sound of the Arduino.