Update and rename ShyTrashcansV10_esp32.ino to ShyTrashcans:ESP32
Added ANGRY states. When you stay with the trashcans for a while, they start getting angry, but opening and closing their lids aggressively. Als added OTA functionality for the ESP32 and added a function where if you send a UDP message to the ESP's IP, it would turn on a light. Everything is working properly.